# openapi/nutrition-api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nutrition Backend API",
    "version": "1.0.0",
    "description": "Owner-approved public API v1 for owner-scoped analysis history and workflow continuation.",
    "x-owner-decision-ref": "docs/OWNER_DECISIONS_V1.md#owner-be-003--product-api-v1",
    "x-production-gate": "docs/OWNER_DECISIONS_V1.md#owner-be-006--production-gate"
  },
  "servers": [{"url": "/"}],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/v1/nutrition/analyses/{analysis_id}/clarifications": {
      "post": {
        "operationId": "answerClarification", "parameters": [{"$ref": "#/components/parameters/AnalysisId"}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClarificationAnswerRequest"}}}},
        "responses": {"200": {"description": "New completed revision.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnalysisSnapshot"}}}}, "400": {"$ref": "#/components/responses/InvalidRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "409": {"$ref": "#/components/responses/StaleClarification"}, "422": {"$ref": "#/components/responses/InsufficientEvidence"}, "500": {"$ref": "#/components/responses/InternalError"}}
      }
    },
    "/v1/nutrition/analyses/{analysis_id}/corrections": {
      "post": {
        "operationId": "correctAnalysis", "parameters": [{"$ref": "#/components/parameters/AnalysisId"}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CorrectionRequest"}}}},
        "responses": {"200": {"description": "New corrected revision.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnalysisSnapshot"}}}}, "400": {"$ref": "#/components/responses/InvalidRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/InsufficientEvidence"}, "500": {"$ref": "#/components/responses/InternalError"}, "503": {"$ref": "#/components/responses/ParserUnavailable"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}},
    "parameters": {
      "AnalysisId": {"name": "analysis_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[\\x20-\\x7E]{1,128}$"}}
    },
    "responses": {
      "InvalidRequest": {"description": "Invalid request or invalid cursor.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Unauthorized": {"description": "Authentication is required.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Forbidden": {"description": "The principal does not own the resource.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "Conflict": {"description": "A stale revision or idempotency conflict.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "StaleClarification": {"description": "The clarification is stale.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "InsufficientEvidence": {"description": "Evidence is insufficient.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "ParserUnavailable": {"description": "The parser is unavailable.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "InternalError": {"description": "An internal error occurred.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}
    },
    "schemas": {
...
      "ErrorEnvelope": {"type": "object", "required": ["error"], "properties": {"error": {"type": "object", "required": ["code", "message"], "properties": {"code": {"type": "string", "enum": ["invalid_request", "invalid_cursor", "unauthorized", "forbidden", "analysis_not_found", "revision_conflict", "stale_clarification", "idempotency_conflict", "analysis_insufficient", "parser_unavailable", "internal_error"]}, "message": {"type": "string"}}, "additionalProperties": false}}, "additionalProperties": false},
...
      "ClarificationAnswerRequest": {"type": "object", "required": ["expected_revision_id", "question_id", "option_id"], "properties": {"expected_revision_id": {"type": "string", "format": "uuid"}, "question_id": {"type": "string", "format": "uuid"}, "option_id": {"type": "string", "minLength": 1}, "mass_g": {"type": ["number", "null"], "exclusiveMinimum": 0}}, "additionalProperties": false},
      "CorrectionRequest": {"type": "object", "required": ["base_revision_id", "item_corrections"], "properties": {"base_revision_id": {"type": "string", "format": "uuid"}, "item_corrections": {"type": "array", "minItems": 1, "maxItems": 10, "items": {"type": "object", "required": ["item_index", "quantity", "unit"], "properties": {"item_index": {"type": "integer", "minimum": 0}, "quantity": {"type": "number", "exclusiveMinimum": 0}, "unit": {"type": "string", "minLength": 1}}, "additionalProperties": false}}}, "additionalProperties": false},
...
      "AnalysisSnapshot": {"type": "object", "required": ["analysis_id", "revision_id", "revision_number", "status", "locale", "versions", "items", "requested_nutrients", "calculation", "is_estimate"], "properties": {"analysis_id": {"type": "string", "format": "uuid"}, "revision_id": {"type": "string", "format": "uuid"}, "revision_number": {"type": "integer", "minimum": 1}, "status": {"type": "string", "enum": ["completed", "insufficient_evidence"]}, "locale": {"type": "string"}, "versions": {"$ref": "#/components/schemas/BehaviorVersions"}, "items": {"type": "array", "items": {"$ref": "#/components/schemas/AnalysisItem"}}, "requested_nutrients": {"type": "array", "items": {"type": "string"}}, "calculation": {"type": "object"}, "is_estimate": {"type": "boolean"}}, "additionalProperties": false},
      "AnalysisItem": {"type": "object", "required": ["source_text", "food_id", "food_name", "profile_id", "estimated_mass_g", "mass_resolution_method", "evidence_quality", "assumptions"], "properties": {"source_text": {"type": "string"}, "food_id": {"type": "string", "format": "uuid"}, "food_name": {"type": "string"}, "quantity": {"type": ["number", "null"]}, "unit_phrase": {"type": ["string", "null"]}, "profile_id": {"type": "string", "format": "uuid"}, "portion_observation_id": {"type": ["string", "null"], "format": "uuid"}, "estimated_mass_g": {"type": "number"}, "lower_mass_g": {"type": ["number", "null"]}, "upper_mass_g": {"type": ["number", "null"]}, "mass_resolution_method": {"type": "string"}, "evidence_quality": {"type": "string"}, "assumptions": {"type": "array", "items": {"type": "string"}}}, "additionalProperties": false},
      "BehaviorVersions": {"type": "object", "required": ["application_version", "parser_schema_version", "prompt_version", "model_provider_version", "normalization_version", "resolution_policy_version", "portion_policy_version", "composition_policy_version", "clarification_policy_version", "correction_policy_version", "calculation_engine_version", "catalog_release_id"], "properties": {"application_version": {"type": "string"}, "parser_schema_version": {"type": "string"}, "prompt_version": {"type": "string"}, "model_provider_version": {"type": "string"}, "normalization_version": {"type": "string"}, "resolution_policy_version": {"type": "string"}, "portion_policy_version": {"type": "string"}, "composition_policy_version": {"type": "string"}, "clarification_policy_version": {"type": "string"}, "correction_policy_version": {"type": "string"}, "calculation_engine_version": {"type": "string"}, "catalog_release_id": {"type": "string", "format": "uuid"}}, "additionalProperties": false}
    }
  }
}